Removed redundant Authentication Library
