+++ src/DwitTech.AccountService.WebApi/Program.cs
@@ -1,4 +1,3 @@
-using DwitTech.Authentication;
 using System.Text.Json.Serialization;
 
 namespace DwitTech.AccountService.WebApi
